switch to correct types for clarity

+++ DerpNES.Tests/BitOperationTests.cs
@@ -10,18 +10,18 @@ public class BitOperationTests
     [Test]
     public void Test_AND()
     {
-        uint a = 0b0110_1110;
-        uint b = 0b0000_1000;
-        uint result = a & b;
+        u8 a = 0b0110_1110;
+        u8 b = 0b0000_1000;
+        u8 result = (u8)(a & b);
         Assert.That( result, Is.EqualTo( 0b0000_1000 ) );
     }
 
     [Test]
     public void Test_XOR()
     {
-        uint a = 0b0110_1110;
-        uint b = 0b0010_0000;
-        uint result = a ^ b;
+        u8 a = 0b0110_1110;
+        u8 b = 0b0010_0000;
+        u8 result = (u8)(a ^ b);
         //         0100_1110
         var test = result.ToBinary();
         Assert.That( result, Is.EqualTo( 0b0100_1110 ) );
@@ -30,9 +30,9 @@ public class BitOperationTests
     [Test]
     public void Test_OR()
     {
-        uint a = 0b0110_1110;
-        uint b = 0b0010_0000;
-        uint result = a | b;
+        u8 a = 0b0110_1110;
+        u8 b = 0b0010_0000;
+        u8 result = (u8)(a | b);
         //         0110_1110
         var test = result.ToBinary();
         Assert.That( result, Is.EqualTo( 0b0110_1110 ) );
@@ -41,8 +41,8 @@ public class BitOperationTests
     [Test]
     public void Test_NOT()
     {
-        uint a = 0b0110_1110;
-        uint result = (uint)(a & ~(1 << 7));
+        u8 a = 0b0110_1110;
+        u8 result = (u8)(a & ~(1 << 7));
         var b = Convert.ToString( result, 2 );
         //         1001_0001
         Assert.That( result, Is.EqualTo( 0b0110_1110 ) );
