+++ DerpNES/Cpu6502.Instructions.cs
@@ -3,7 +3,7 @@
 // https://masswerk.at/6502/6502_instruction_set.html#LSR
 public partial class Cpu6502
 {
-    uint Illegal()
+    u8 Illegal()
     {
         throw new NotImplementedException();
     }
@@ -12,7 +12,7 @@ public partial class Cpu6502
     /// Loads a byte of memory into the accumulator
     /// </summary>
     /// <returns></returns>
-    uint LDA()
+    u8 LDA()
     {
         A = NextByte();
         SetFlag( StatusFlag.Zero, A == 0 );
@@ -24,13 +24,13 @@ public partial class Cpu6502
     /// <summary>
     /// Add with carry
     /// </summary>
-    uint ADC() => throw new NotImplementedException();
+    u8 ADC() => throw new NotImplementedException();
 
     /// <summary>
     /// AND Memory with Accumulator
     /// </summary>
     /// <returns></returns>
-    uint AND()
+    u8 AND()
     {
         A &= FetchData();
         SetFlag(StatusFlag.Zero, A == 0 );
@@ -40,7 +40,7 @@ public partial class Cpu6502
         return 1; // additional clock cycle
     }
 
-    uint BRK() => 0;
+    u8 BRK() => 0;
 
-    uint ORA() => 0;
+    u8 ORA() => 0;
 }
